adding data jpa to spring for h2

+++ src/main/java/com/in28minutes/rest/webservices/restfulwebservices/todo/TodoHardcodedService.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Service;
 public class TodoHardcodedService {
 
 	private static List<Todo> todos = new ArrayList<Todo>();
-	private static int idCounter=0;
+	private static Long idCounter=0l;
 	static {
 		todos.add(new Todo(++idCounter,"in28minutes","Learn to dance",new Date(),false));
 		todos.add(new Todo(++idCounter,"in28minutes","Learn about Microservice",new Date(),false));
